Conflicts:
	src/server/scripts/Pandaria/HeartOfFear/heart_of_fear.cpp

+++ src/server/game/Spells/SpellScript.cpp
@@ -80,14 +80,14 @@ uint32 _SpellScript::EffectHook::GetAffectedEffectsMask(SpellInfo const* spellEn
     else
     {
         if (CheckEffect(spellEntry, effIndex))
-            mask |= (uint8)1<<effIndex;
+            mask |= 1<<effIndex;
     }
     return mask;
 }
 
 bool _SpellScript::EffectHook::IsEffectAffected(SpellInfo const* spellEntry, uint8 effIndex)
 {
-    return GetAffectedEffectsMask(spellEntry) & 1<<effIndex;
+    return GetAffectedEffectsMask(spellEntry) & ((uint32)1 << (uint32)effIndex);
 }
 
 std::string _SpellScript::EffectHook::EffIndexToString()
